requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
pymongo==4.6.1
python-dotenv==1.0.0
dnspython==2.4.2
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
colorama==0.4.6
packaging==23.2
rich==13.7.0
ordered-set==4.1.0
typing-extensions==4.9.0
importlib-resources==6.1.1
redis==5.0.1
limits==3.7.0
Deprecated==1.2.14
wrapt==1.16.0
pytz==2023.3
